Modify rc data storage unit tests. use Mocked DB(dict type) instead of levelDB
